+++ src/OpenTelemetry.Exporter.Prometheus.AspNetCore/CHANGELOG.md
@@ -7,6 +7,10 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* **Breaking Change** Removed `DisableTimestamp` property from
+  `PrometheusAspNetCoreOptions`.
+  ([#7176](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7176))
+
 ## 1.15.3-beta.1
 
 Released 2026-Apr-21